revert move of test_period_asfreq
